+++ src/features/permissions/permissions.ts
@@ -1,0 +1,3 @@
+export { default as Permission } from "./components/Permission/Permission";
+export * from "./logics/permissionLogic";
+export * from "./constants/permissionScopes";
