+++ src/routes/Routes.tsx
@@ -1,8 +1,9 @@
 import React, { FC, memo } from "react";
 import { Switch } from "react-router-dom";
-import HomeScreen from "features/home/screens/HomeScreen/HomeScreen";
-import { setPermissions } from "components/Permission/Permission";
-import { PermissionEnum } from "constants/permissionScopes";
+import {
+  setPermissions,
+  PermissionEnum,
+} from "features/permissions/permissions";
 import RouteWrapper from "./RouteWrapper";
 import routeList from "./Routes.config";
 
@@ -15,9 +16,6 @@ const Routes: FC = () => {
       {routeList.map(route => (
         <RouteWrapper key={route.path} {...route} />
       ))}
-
-      {/* redirect user to Home page if route does not exist and user is not authenticated */}
-      <RouteWrapper component={HomeScreen} />
     </Switch>
   );
 };
